+++ commons/src/main/java/org/fuin/srcgen4j/commons/SrcGen4JConfig.java
@@ -66,6 +66,12 @@ public class SrcGen4JConfig {
 
     @Nullable
     @Valid
+    @XmlElementWrapper(name = "replacers")
+    @XmlElement(name = "replacer")
+    private List<Replacer> replacers;
+
+    @Nullable
+    @Valid
     @XmlElement(name = "parsers")
     private Parsers parsers;
 
@@ -136,6 +142,26 @@ public class SrcGen4JConfig {
      */
     public final void setProjects(@Nullable final List<Project> projects) {
         this.projects = projects;
+    }
+
+    /**
+     * Returns a list of replacers.
+     *
+     * @return Replacers.
+     */
+    @Nullable
+    public final List<Replacer> getReplacers() {
+        return replacers;
+    }
+
+    /**
+     * Sets a list of replacers.
+     *
+     * @param replacers
+     *            Replacers.
+     */
+    public final void setReplacers(@Nullable final List<Replacer> replacers) {
+        this.replacers = replacers;
     }
 
     /**
@@ -244,6 +270,11 @@ public class SrcGen4JConfig {
         if (projects != null) {
             for (final Project project : projects) {
                 project.init(context, this, varMap);
+            }
+        }
+        if (replacers != null) {
+            for (final Replacer replacer : replacers) {
+                replacer.init(varMap);
             }
         }
         if (generators != null) {