+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release notes for srcgen4j.
 - Uses **Java 17**
 - Switched to `jakarta` namespace
 - Use fuin.org BOM
+- Added [Replacer](commons/src/main/java/org/fuin/srcgen4j/commons/Replacer.java)
 
 ### [0.4.2]
 - Uses **Java 8**